refactor: remove constants and coalesce to utils
