+++ player.js
@@ -1,5 +1,4 @@
-import { SpeedTypes } from "./constants.js";
-import Utils from "./utils.js";
+import { Utils, SpeedTypes } from "./utils.js";
 
 export default class Player {
     constructor(scene, x, y) {
